requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
googlesearch-python>=1.1.0